Refactoring start-build.sh

+++ start-build.sh
@@ -1,4 +1,4 @@
-#!/bin/sh
+#!/bin/sh -e
 
 # Ensure that we have superpowers.
 if [ $(id -u) != 0 ]; then
@@ -6,8 +6,10 @@ if [ $(id -u) != 0 ]; then
         exit
 fi
 
-docker rm -f CodingPiratesOS
-set -e
+if [ $(docker ps -a -f "name=/CodingPiratesOS$" -q) ]; then
+  docker rm -f CodingPiratesOS
+fi
+
 docker build -t darkmagus/codingpiratesos .
 
 docker run -t --privileged -v $PWD/build:/usr/app/build -v $PWD/images:/usr/app/images -v $PWD/basics:/usr/app/basics -v $PWD/data:/usr/app/data -v $PWD/scripts:/usr/app/scripts --name="CodingPiratesOS" darkmagus/codingpiratesos
